refactor : 당첨 번호 생성 시 필요한 기능을 WinLotto 클래스로 분리
